group response changes

+++ TOMI.Services/Repository/GroupRepository.cs
@@ -26,38 +26,55 @@ namespace TOMI.Services.Repository
             _mapper = mapper;
         }
 
-        public async Task<GroupModel> AddGroup(GroupModel model)
+        public async Task<GroupResponse> AddGroup(GroupModel model)
         {
             Group existingRanges = await _context.Group.FirstOrDefaultAsync(c => c.Id == model.Id);
+            if (existingRanges != null)
+            {
+                var isNameExist = _context.Group.FirstOrDefaultAsync(c => c.Name == model.Name);
 
-            var group = _mapper.Map<Group>(model);
-            if (existingRanges == null)
-            {
-                Group result = _context.Group.Add(group).Entity;
-            }
-            else
-            {
+                if (isNameExist != null)
+                    return new GroupResponse { Error = "Group name already exist ", Success = false };
+
                 existingRanges.Name = model.Name;
                 _context.Group.Update(existingRanges);
+                _context.SaveChanges();
+                return new GroupResponse { group = existingRanges, Success = true };
             }
-            _context.SaveChanges();
+    
+            if (existingRanges == null)
+            {
+                var group = _mapper.Map<Group>(model);
+                Group result = _context.Group.Add(group).Entity;
+                _context.SaveChanges();
+                return new GroupResponse { group = result, Success = true };
+            }
+      
 
-            return model;
             throw new ValidationException("Group not found!");
         }
 
        
-        public async Task<Group> DeleteGroup(Guid id)
+        public async Task<GroupResponse> DeleteGroup(Guid id)
         {
             var result= await _context.Group
                 .FirstOrDefaultAsync(e => e.Id == id);
             if (result != null)
             {
+                var recordexist = await _context.Group
+                 .FirstOrDefaultAsync(e => e.Id == id);
+                if (recordexist != null)
+                    return new GroupResponse {Error = "Group is used in ranges ",Success = false};
+
                 _context.Group.Remove(result);
+
                 await _context.SaveChangesAsync();
-                return result;
+
+                return new GroupResponse { group = result,Success = true};
             }
-            return null;
+            else
+             return new GroupResponse {Error = "Group don't exist",Success = false};
+            
         }
 
         public async Task<Group> GetGroup(Guid id)
